refactor websocket handler
