+++ package.json
@@ -27,6 +27,7 @@
     "@types/ws": "^7.4.7",
     "prettier": "^2.3.2",
     "prettier-plugin-svelte": "^2.4.0",
+    "strict-event-emitter-types": "^2.0.0",
     "svelte": "^3.38.3",
     "svelte-preprocess-esbuild": "^2.0.0",
     "tslib": "^2.3.1",
