+++ src/lib/server/ApiServer.test.ts
@@ -8,8 +8,6 @@ import {ApiServer} from '$lib/server/ApiServer.js';
 import {Database} from '$lib/db/Database.js';
 import {default_postgres_options} from '$lib/db/postgres.js';
 import {WebsocketServer} from '$lib/server/WebsocketServer.js';
-import {to_handle_websocket_message} from '$lib/server/handle_websocket_message';
-import {services} from '$lib/server/services';
 
 const TEST_PORT = 3003; // TODO
 
@@ -18,13 +16,10 @@ const test__api_server = suite('ApiServer');
 
 test__api_server('init and close', async () => {
 	const server = createServer();
-	const handle_websocket_message = to_handle_websocket_message(services);
 	const api_server: ApiServer = new ApiServer({
 		server,
 		app: polka({server}),
-		websocket_server: new WebsocketServer(server, (websocket_server, socket, message, account_id) =>
-			handle_websocket_message(api_server, websocket_server, socket, message, account_id),
-		),
+		websocket_server: new WebsocketServer(server),
 		db: new Database({sql: postgres(default_postgres_options)}),
 		port: TEST_PORT,
 	});
